requirements.txt
-e git+https://github.com/sgarciaest/albcovis-project.git@fff5387b74218a9b257776ffe67074ed19171323#egg=albcovis
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorthief==0.2.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dlib==20.0.0
executing==2.2.0
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.2
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
